Refactoring process 1
We have moved the old codes to archive location and refactored the initial codes. Modularized the python scripts by implementing functions. We have introduced Configuration file config.csv which stores the code parameters as a best practice followed in the Industry. This config.csv file will used by all the scripts and hard coding of values will be avoided.

Co-Authored-By: sanv00 <45181379+sanv00@users.noreply.github.com>